[app:contexts] Rename dir `app/context/` to `app/contexts`
[app:App] Update reference `app/context` -> `app/contexts`
[app:Head] Update reference `app/context` -> `app/contexts`
[app:HTMLBody] Update reference `app/context` -> `app/contexts`
