+++ app/Head/__tests__/index.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 
-import ManifestContext from 'app/context/ManifestContext';
+import ManifestContext from 'app/contexts/ManifestContext';
 import useRouteHead from 'app/hooks/useRouteHead';
 
 import Head from '../index';
