+++ app/contexts/__tests__/ManifestContext.test.js
